pipeline/auditlog/src/requirements.txt
elasticsearch
google-cloud-pubsub